Add support for splits
Added support for adding splits to a transaction. Transactions either
accept a single debtor directly on the transaction, or it takes one or
more splits.

Splits can be either amounts or shares. The sum of all shares need to be
equal to the share denominator. Validation is performed after splits are
added to a transaction.

Added transaction tests with splits.

Also in this change:

- Renamed group.users to group.members
- Fixed circular import issues